+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTS4FJFEW4FQEBPHGFC5GXEQ"
+        "01KTTH7ERBEAPDPYA3VCAMTGQX"
     )
     r"""The unique identifier of the data source"""