+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_feedback_evaluationop.py
@@ -55,7 +55,7 @@ class RequestBody3(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodyFeedbackSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-10T16:05:09.960Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-11T05:07:09.910Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -121,7 +121,7 @@ class PostV2FeedbackEvaluationRequestBody2(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-10T16:05:09.960Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-11T05:07:09.909Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -187,7 +187,7 @@ class PostV2FeedbackEvaluationRequestBody1(BaseModel):
 
     source: Optional[RequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-10T16:05:09.959Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-11T05:07:09.908Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")